+++ src/app/app-routing.module.ts
@@ -3,7 +3,7 @@ import { Routes, RouterModule, ExtraOptions } from '@angular/router';
 import { AboutComponent, PathNotFoundComponent } from './layout';
 import { CartComponent } from './feature/cart/cart.component';
 import { CustomPreloadingStrategyService } from './feature/core/custom-preloading-strategy.service';
-import { FeedbacksComponent } from './feature/product/components';
+import { OrderFormComponent } from './feature/order/components/order-form/order-form.component';
 
 const extraOptions: ExtraOptions = {
   preloadingStrategy: CustomPreloadingStrategyService,
@@ -27,6 +27,10 @@ const routes: Routes = [
       preload: true
     }
   }, 
+  {
+    path: 'order',
+    component: OrderFormComponent
+  },
   {
     // The router will match this route if the URL requested
     // doesn't match any paths for routes defined in our configuration