+++ src/app/feature/order/order.module.ts
@@ -4,10 +4,12 @@ import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 import { SharedModule } from 'src/app/shared/shared.module';
 import { OrderFormComponent } from '../order/components/order-form/order-form.component';
+import { OrderService } from './services/order.service';
+import { CoreModule } from 'src/app/core/core.module';
 @NgModule({
-    imports: [CommonModule, FormsModule, SharedModule],
+    imports: [CommonModule, FormsModule, SharedModule,CoreModule],
     declarations: [OrderFormComponent],
-    providers: [],
+    providers: [OrderService],
     exports: [OrderFormComponent]
 })
 export class OrderModule { }